feat(widgets): Layout, StickyChatWidget widgets 레이어로 분리
components에는 re-export 유지로 기존 임포트 경로 호환 보장
