+++ src/widgets/StickyChatWidget/index.ts
@@ -1,0 +1,1 @@
+export { StickyChatWidget } from './StickyChatWidget'
